Host Test
Added a third host to the network and conducted messaging test.
  1
 / \
0 - 2
Links were conncted in a unidirectional way (cw), which meant if
host 0 wanted to send a message to host 2, it woul have to send it
to host 1 first. However host 1 will reject the message.

Need to test this again with more pipes to create a bidirectional
link... Added rcvflag as an int to clear an error. Base file was
precompiled, so the error did not come up until a force make was
done.